feature(car): Admin Can get all cars whether sold or unsold - implement feature - Delivers [#166790946]
